projeto-de-programas: utiliza bscpp em todos os exemplos

+++ 06-projeto-de-programas/exemplos/numero_azulejos.cpp
@@ -3,12 +3,25 @@
 
 using namespace std;
 
+// Análise
+//
+// Calcular o número de azulejos necessários para azulejar uma parede com
+// determinado comprimento e altura,  considerando que o azulejo mede 0,2,0m x
+// 0,2m e que nenhum azulejo é perdido e que recordes são
+// descartados.
+//
+// Tipos de dados
+//
+// O comprimento e a altura da parede são dados em metros e representados com
+// números positivos. O número de azulejos é representado por um número inteiro
+// positivo.
+
 // Calcula o número de azulejos de 0,2mx0,2m necessários para azulejar uma area
 // de tamanho comprimento x altura (em metros) considerando que nenhum azulejo
 // é perdido e que recordes são descartados.
 int numero_azulejos(double comprimento, double altura)
 {
-    return ceil(comprimento / 0.2) * ceil(altura / 0.2);
+    return (int)(ceil(comprimento / 0.2) * ceil(altura / 0.2));
 }
 
 examples
